requirements.txt
requests
sqlalchemy
cachetools
python-dotenv